fixed Propay utility added utility to caseService changed some datatypes

+++ src/main/java/de/propra2/ausleiherino24/model/Article.java
@@ -33,9 +33,9 @@ public class Article {
 
     private String image;
 
-    private int deposit;
+    private Double deposit;
 
-    private int costPerDay;
+    private Double costPerDay;
 
     private String location;
